requirements.txt
torch
torchvision
numpy
pandas
matplotlib
scikit-learn
PyYAML
Pillow
tqdm
fastapi
uvicorn
python-multipart